Refactor platformCoroutineDispatcher
